rename ImageManager to FileManager
